Revert "UI and text changes in avatar screen dialogue box"
This reverts commit a24339f50cf25e1c229ef133c9bb42e1ab023683.
