+++ public/locales/en/translation.json
@@ -333,7 +333,6 @@
   "Place Value 300-500": "Place Value 300-500",
   "Play ' x1 ' to win your daily reward": "Play ' x1 ' to win your daily reward",
   "Play ' x1 ' to win your weekly reward": "Play ' x1 ' to win your weekly reward",
-  "Playing for just ' x1 ' every week helps you do better in school and get ready for more learning!": "Playing for just ' x1 ' every week helps you do better in school and get ready for more learning!",
   "Please create your child's profile.": "Please create your child's profile.",
   "Please Enter 10 digit Mobile Number": "Please Enter 10 digit Mobile Number",
   "Please Enter 6 Digit Code": "Please Enter 6 Digit Code",
