+++ public/locales/hi/translation.json
@@ -342,7 +342,6 @@
   "Place Value 300-500": "स्थानीय मान 300 -500",
   "Play ' x1 ' to win your daily reward": "अपना दैनिक इनाम जीतने के लिए ' x1 ' खेलें",
   "Play ' x1 ' to win your weekly reward": "अपना साप्ताहिक इनाम जीतने के लिए ' x1 ' खेलें।",
-  "Playing for just ' x1 ' every week helps you do better in school and get ready for more learning!": "हर हफ्ते सिर्फ ' x1 ' खेलने से आपको स्कूल में बेहतर प्रदर्शन करने और अधिक सीखने के लिए तैयार होने में मदद मिलती है!",
   "Please create your child's profile.": "कृपया अपने बच्चे की प्रोफ़ाइल बनाएं.",
   "Please Enter 10 digit Mobile Number": "कृपया 10 अंकों का मोबाइल नंबर दर्ज करें।",
   "Please Enter 6 Digit Code": "कृपया 6-अंकीय कोड दर्ज करें।",
