+++ public/locales/kn/translation.json
@@ -338,7 +338,6 @@
   "Place Value 300-500": "ಸ್ಥಳದ ಮೌಲ್ಯ 300 - 500",
   "Play ' x1 ' to win your daily reward": "ನಿಮ್ಮ ದೈನಂದಿನ ಬಹುಮಾನವನ್ನು ಗೆಲ್ಲಲು 'x1' ಪ್ಲೇ ಮಾಡಿ",
   "Play ' x1 ' to win your weekly reward": "ನಿಮ್ಮ ಸಾಪ್ತಾಹಿಕ ಬಹುಮಾನವನ್ನು ಗೆಲ್ಲಲು 'x1' ಆಟವಾಡಿ.",
-  "Playing for just ' x1 ' every week helps you do better in school and get ready for more learning!": "ಪ್ರತಿ ವಾರ ಕೇವಲ ' x1 ' ಆಡುವುದರಿಂದ ಶಾಲೆಯಲ್ಲಿ ಉತ್ತಮವಾಗಿ ಪ್ರದರ್ಶಿಸಲು ಮತ್ತು ಹೆಚ್ಚಿನ ಕಲಿಕೆಗೆ ಸಹಾಯವಾಗುತ್ತದೆ.",
   "Please create your child's profile.": "ದಯವಿಟ್ಟು ನಿಮ್ಮ ಮಗುವಿನ ಪ್ರೊಫೈಲ್ ಅನ್ನು ರಚಿಸಿ",
   "Please Enter 10 digit Mobile Number": "ದಯವಿಟ್ಟು 10 ಅಂಕಿಯ ಮೊಬೈಲ್ ಸಂಖ್ಯೆಯನ್ನು ನಮೂದಿಸಿ",
   "Please Enter 6 Digit Code": "ದಯವಿಟ್ಟು 6 ಅಂಕಿ ಕೋಡ್ ನಮೂದಿಸಿ",
