+++ src/components/animation/Avatar.tsx
@@ -56,13 +56,12 @@ export class AvatarObj {
   currentLesson: Lesson | undefined;
   currentLessonSuggestionIndex: number;
   currentRecommendedLessonIndex: number = 0;
-  weeklyProgressGoal: number = 10;
+  weeklyProgressGoal: number = 25;
   weeklyTimeSpent: {} = { min: 0, sec: 0 };
   weeklyPlayedLesson: number = 0;
   wrongAttempts: number = 0;
   // gamifyTimespentMessage = "Play ' x1 ' to win your daily reward";
-  gamifyTimespentMessage =
-    "Playing for just ' x1 ' every week helps you do better in school and get ready for more learning!";
+  gamifyTimespentMessage = "Play ' x1 ' to win your weekly reward";
 
   private constructor() {}
 
@@ -202,7 +201,7 @@ export class AvatarObj {
       } else if (showDailyProgress === "true") {
         if (this.weeklyTimeSpent["min"] * 60 < this.weeklyProgressGoal * 60) {
           await this.loadAvatarWeeklyProgressData();
-          return;
+          return
         }
 
         if (this.weeklyTimeSpent["min"] * 60 >= this.weeklyProgressGoal * 60) {
@@ -276,6 +275,7 @@ export class AvatarObj {
   }
 
   public async loadAvatarNextSuggestion() {
+    
     if (this._currentSuggestionNumber === this._allSuggestions.length - 1) {
       this._currentSuggestionNumber = 0;
       localStorage.setItem(
@@ -308,6 +308,7 @@ export class AvatarObj {
 
   public async loadAvatarWeeklyProgressData() {
     try {
+
       const currentStudent = await Util.getCurrentStudent();
       if (!currentStudent) {
         return;
