+++ src/utility/util.ts
@@ -81,7 +81,6 @@ import { Router } from "react-router-dom";
 import { schoolUtil } from "./schoolUtil";
 import lesson from "../models/lesson";
 import Lesson from "../models/lesson";
-import Sticker from "../models/Sticker";
 
 declare global {
   interface Window {
@@ -1359,28 +1358,6 @@ export class Util {
 
       throw error;
     }
-  }
-
-  // const getNextUnlockStickers = async (): Promise<(Sticker | undefined)[]> => {
-  public static async getNextUnlockStickers(): Promise<
-    (Sticker | undefined)[]
-  > {
-    const date = new Date();
-    const api = ServiceConfig.getI().apiHandler;
-    const rewardsDoc = await api.getRewardsById(date.getFullYear().toString());
-    if (!rewardsDoc) return [];
-    const currentWeek = Util.getCurrentWeekNumber();
-    const stickerIds: string[] = [];
-    const weeklyData = rewardsDoc.weeklySticker;
-    weeklyData[currentWeek.toString()].forEach((value) => {
-      if (value.type === LeaderboardRewardsType.STICKER) {
-        stickerIds.push(value.id);
-      }
-    });
-    const stickerDocs = await Promise.all(
-      stickerIds.map((value) => api.getStickerById(value))
-    );
-    return stickerDocs;
   }
 
   public static getCurrentWeekNumber() {
